% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.map.R
\name{sheep.map}
\alias{sheep.map}
\title{Create dynamic Shiny maps of most recent sheep locations}
\usage{
sheep.map(vecpath, sheepdb, idcol, nlocs, xcol, ycol, proj)
}
\arguments{
\item{vecpath}{path to where GPS data is stored}

\item{sheepdb}{path to where capture database is stored}

\item{idcol}{name of column that should be subset (AID, CollarSerial, etc)}

\item{nlocs}{number of locations to include in map}

\item{xcol}{name of column where x coordinates are stored}

\item{ycol}{name of column where y coordinates are stored}

\item{proj}{character string of projection for coordinates}
}
\value{
Returns a raster brick with all climate data for specific year
}
\description{
Create dynamic Shiny maps of most recent sheep locations
}
\examples{

}
\keyword{climate,}
\keyword{pdsi,}
\keyword{swe}
\keyword{temperature,}
